+++ code/packages/rust/spice-netlist-parser/src/lib.rs
@@ -786,6 +786,7 @@ fn parse_element(
                     .or_else(|| model.params.get("CBC"))
                     .unwrap_or(&0.0),
                 *model.params.get("TF").unwrap_or(&0.0),
+                *model.params.get("TR").unwrap_or(&0.0),
             )))
         }
         'J' => {